Conflicts:
	app/Console/cake.php
	lib/Cake/Console/ShellDispatcher.php
	lib/Cake/Console/Templates/skel/Console/cake.php
	lib/Cake/Console/cake.php
	lib/Cake/Controller/Scaffold.php
	lib/Cake/TestSuite/Fixture/CakeFixtureManager.php
	lib/Cake/View/Errors/missing_layout.ctp
	lib/Cake/View/Errors/missing_view.ctp
	src/Auth/BasicAuthenticate.php
	src/Utility/String.php
	src/View/View.php

+++ src/TestSuite/Fixture/FixtureManager.php
@@ -248,7 +248,9 @@ class FixtureManager {
 						if (!in_array($db->configName(), (array)$fixture->created)) {
 							$this->_setupTable($fixture, $db, $test->dropTables);
 						}
-						$fixture->truncate($db);
+						if (!$test->dropTables) {
+							$fixture->truncate($db);
+						}
 						$fixture->insert($db);
 					}
 				});
@@ -299,7 +301,9 @@ class FixtureManager {
 			if (!in_array($db->configName(), (array)$fixture->created)) {
 				$this->_setupTable($fixture, $db, $dropTables);
 			}
-			$fixture->truncate($db);
+			if (!$dropTables) {
+				$fixture->truncate($db);
+			}
 			$fixture->insert($db);
 		} else {
 			throw new \UnexpectedValueException(sprintf('Referenced fixture class %s not found', $name));
